// src/client/api/generated/zod/spec/spec.zod.ts
export const createSpecBodyFileMax = 1024;

export const createSpecBodyContentDefault = ``;
export const createSpecBodyContentMax = 5000000;



export const CreateSpecBody = zod.object({
  "file": zod.string().min(1).max(createSpecBodyFileMax).describe('Tên file .md tương đối trong thư mục spec (vd "sidebar.md", "ui/header.md")'),
  "content": zod.string().max(createSpecBodyContentMax).default(createSpecBodyContentDefault)
})
// ...
export const renameSpecBodyNewFileMax = 1024;



export const RenameSpecBody = zod.object({
  "newFile": zod.string().min(1).max(renameSpecBodyNewFileMax).describe('Tên file .md tương đối trong thư mục spec (vd "sidebar.md", "ui/header.md")')
})